Divided the project into separate packages to make it easier to read. Changed the former checkOutBook method from Book class to checkInCheckOutBook and now it gets one more argument. It is now possible to check books in and out by modyfing the availability column in the db. Some of the comments have been changed to reflect new functionality and new plans
