+++ src/User.java
@@ -1,0 +1,3 @@
+class User {
+
+}
